Detach Policies and add PoliciesMixin
