+++ gvm/protocols/gmpv208/__init__.py
@@ -52,6 +52,7 @@ from gvm.protocols.gmpv208.entities.operating_systems import (
     OperatingSystemsMixin,
 )
 from gvm.protocols.gmpv208.entities.overrides import OverridesMixin
+from gvm.protocols.gmpv208.entities.policies import PoliciesMixin
 from gvm.protocols.gmpv208.entities.results import ResultsMixin
 from gvm.protocols.gmpv208.entities.report_formats import (
     ReportFormatType,
@@ -90,6 +91,7 @@ class Gmp(
     NotesMixin,
     OperatingSystemsMixin,
     OverridesMixin,
+    PoliciesMixin,
     PortListMixin,
     ReportsMixin,
     ResultsMixin,
