+++ gvm/protocols/gmpv208/gmpv208.py
@@ -52,7 +52,6 @@ from . import types
 from .types import *  # pylint: disable=unused-wildcard-import, wildcard-import
 from .types import _UsageType as UsageType
 
-_EMPTY_POLICY_ID = '085569ce-73ed-11df-83c3-002264764cea'
 
 PROTOCOL_VERSION = (20, 8)
 
@@ -271,30 +270,6 @@ class GmpV208Mixin(GvmProtocol):
             _xmlresource.add_element("type", _actual_resource_type.value)
 
         return self._send_xml_command(cmd)
-
-    def create_policy(
-        self, name: str, *, policy_id: str = None, comment: Optional[str] = None
-    ) -> Any:
-        """Create a new policy config
-
-        Arguments:
-            name: Name of the new policy
-            policy_id: UUID of an existing policy as base. By default the empty
-                policy is used.
-            comment: A comment on the policy
-
-        Returns:
-            The response. See :py:meth:`send_command` for details.
-        """
-        if policy_id is None:
-            policy_id = _EMPTY_POLICY_ID
-        return self.__create_config(
-            config_id=policy_id,
-            name=name,
-            comment=comment,
-            usage_type=UsageType.POLICY,
-            function=self.create_policy.__name__,
-        )
 
     def create_tag(
         self,
@@ -635,106 +610,6 @@ class GmpV208Mixin(GvmProtocol):
 
         return self._send_xml_command(cmd)
 
-    def modify_policy_set_nvt_preference(
-        self,
-        policy_id: str,
-        name: str,
-        nvt_oid: str,
-        *,
-        value: Optional[str] = None,
-    ) -> Any:
-        """Modifies the nvt preferences of an existing policy.
-
-        Arguments:
-            policy_id: UUID of policy to modify.
-            name: Name for preference to change.
-            nvt_oid: OID of the NVT associated with preference to modify
-            value: New value for the preference. None to delete the preference
-                and to use the default instead.
-        """
-        self.modify_config_set_nvt_preference(
-            config_id=policy_id, name=name, nvt_oid=nvt_oid, value=value
-        )
-
-    def modify_policy_set_name(self, policy_id: str, name: str) -> Any:
-        """Modifies the name of an existing policy
-
-        Arguments:
-            config_id: UUID of policy to modify.
-            name: New name for the config.
-        """
-        self.modify_config_set_name(config_id=policy_id, name=name)
-
-    def modify_policy_set_comment(
-        self, policy_id: str, comment: Optional[str] = ""
-    ) -> Any:
-        """Modifies the comment of an existing policy
-
-        Arguments:
-            policy_id: UUID of policy to modify.
-            comment: Comment to set on a config. Default: ''
-        """
-        self.modify_config_set_comment(config_id=policy_id, comment=comment)
-
-    def modify_policy_set_scanner_preference(
-        self, policy_id: str, name: str, *, value: Optional[str] = None
-    ) -> Any:
-        """Modifies the scanner preferences of an existing policy
-
-        Arguments:
-            policy_id: UUID of policy to modify.
-            name: Name of the scanner preference to change
-            value: New value for the preference. None to delete the preference
-                and to use the default instead.
-
-        """
-        self.modify_config_set_scanner_preference(
-            config_id=policy_id, name=name, value=value
-        )
-
-    def modify_policy_set_nvt_selection(
-        self, policy_id: str, family: str, nvt_oids: List[str]
-    ) -> Any:
-        """Modifies the selected nvts of an existing policy
-
-        The manager updates the given family in the config to include only the
-        given NVTs.
-
-        Arguments:
-            policy_id: UUID of policy to modify.
-            family: Name of the NVT family to include NVTs from
-            nvt_oids: List of NVTs to select for the family.
-        """
-        self.modify_config_set_nvt_selection(
-            config_id=policy_id, family=family, nvt_oids=nvt_oids
-        )
-
-    def modify_policy_set_family_selection(
-        self,
-        policy_id: str,
-        families: List[Tuple[str, bool, bool]],
-        *,
-        auto_add_new_families: Optional[bool] = True,
-    ) -> Any:
-        """
-        Selected the NVTs of a policy at a family level.
-
-        Arguments:
-            policy_id: UUID of policy to modify.
-            families: A list of tuples with the first entry being the name
-                of the NVT family selected, second entry a boolean indicating
-                whether new NVTs should be added to the family automatically,
-                and third entry a boolean indicating whether all nvts from
-                the family should be included.
-            auto_add_new_families: Whether new families should be added to the
-                policy automatically. Default: True.
-        """
-        self.modify_config_set_family_selection(
-            config_id=policy_id,
-            families=families,
-            auto_add_new_families=auto_add_new_families,
-        )
-
     def modify_tag(
         self,
         tag_id: str,
@@ -869,55 +744,27 @@ class GmpV208Mixin(GvmProtocol):
         Returns:
             The response. See :py:meth:`send_command` for details.
         """
-        return self.__get_configs(
-            UsageType.SCAN,
-            filter=filter,
-            filter_id=filter_id,
-            trash=trash,
-            details=details,
-            families=families,
-            preferences=preferences,
-            tasks=tasks,
-        )
+        cmd = XmlCommand("get_configs")
+        cmd.set_attribute("usage_type", "scan")
 
-    def get_policies(
-        self,
-        *,
-        audits: Optional[bool] = None,
-        filter: Optional[str] = None,
-        filter_id: Optional[str] = None,
-        details: Optional[bool] = None,
-        families: Optional[bool] = None,
-        preferences: Optional[bool] = None,
-        trash: Optional[bool] = None,
-    ) -> Any:
-        """Request a list of policies
+        add_filter(cmd, filter, filter_id)
 
-        Arguments:
-            audits: Whether to get audits using the policy
-            filter: Filter term to use for the query
-            filter_id: UUID of an existing filter to use for the query
-            details: Whether to get  families, preferences, nvt selectors
-                and tasks.
-            families: Whether to include the families if no details are
-                requested
-            preferences: Whether to include the preferences if no details are
-                requested
-            trash: Whether to get the trashcan audits instead
+        if trash is not None:
+            cmd.set_attribute("trash", to_bool(trash))
 
-        Returns:
-            The response. See :py:meth:`send_command` for details.
-        """
-        return self.__get_configs(
-            UsageType.POLICY,
-            filter=filter,
-            filter_id=filter_id,
-            details=details,
-            families=families,
-            preferences=preferences,
-            tasks=audits,
-            trash=trash,
-        )
+        if details is not None:
+            cmd.set_attribute("details", to_bool(details))
+
+        if families is not None:
+            cmd.set_attribute("families", to_bool(families))
+
+        if preferences is not None:
+            cmd.set_attribute("preferences", to_bool(preferences))
+
+        if tasks is not None:
+            cmd.set_attribute("tasks", to_bool(tasks))
+
+        return self._send_xml_command(cmd)
 
     def get_config(
         self, config_id: str, *, tasks: Optional[bool] = None
@@ -931,58 +778,21 @@ class GmpV208Mixin(GvmProtocol):
         Returns:
             The response. See :py:meth:`send_command` for details.
         """
-        return self.__get_config(
-            config_id=config_id, usage_type=UsageType.SCAN, tasks=tasks
-        )
-
-    def get_policy(
-        self, policy_id: str, *, audits: Optional[bool] = None
-    ) -> Any:
-        """Request a single policy
-
-        Arguments:
-            policy_id: UUID of an existing policy
-            audits: Whether to get audits using this config
-
-        Returns:
-            The response. See :py:meth:`send_command` for details.
-        """
-        return self.__get_config(policy_id, UsageType.POLICY, tasks=audits)
-
-    def clone_policy(self, policy_id: str) -> Any:
-        """Clone a policy from an existing one
-
-        Arguments:
-            policy_id: UUID of the existing policy
-
-        Returns:
-            The response. See :py:meth:`send_command` for details.
-        """
-        if not policy_id:
+        if not config_id:
             raise RequiredArgument(
-                function=self.clone_policy.__name__, argument='policy_id'
+                function=self.get_config.__name__, argument='config_id'
             )
 
-        cmd = XmlCommand("create_config")
-        cmd.add_element("copy", policy_id)
-        return self._send_xml_command(cmd)
+        cmd = XmlCommand("get_configs")
+        cmd.set_attribute("config_id", config_id)
 
-    def delete_policy(
-        self, policy_id: str, *, ultimate: Optional[bool] = False
-    ) -> Any:
-        """Deletes an existing policy
-        Arguments:
-            policy_id: UUID of the policy to be deleted.
-            ultimate: Whether to remove entirely, or to the trashcan.
-        """
-        if not policy_id:
-            raise RequiredArgument(
-                function=self.delete_policy.__name__, argument='policy_id'
-            )
+        cmd.set_attribute("usage_type", "scan")
 
-        cmd = XmlCommand("delete_config")
-        cmd.set_attribute("config_id", policy_id)
-        cmd.set_attribute("ultimate", to_bool(ultimate))
+        if tasks is not None:
+            cmd.set_attribute("tasks", to_bool(tasks))
+
+        # for single entity always request all details
+        cmd.set_attribute("details", "1")
 
         return self._send_xml_command(cmd)
 
@@ -1030,65 +840,6 @@ class GmpV208Mixin(GvmProtocol):
         cmd.add_element("scanner", scanner_id)
         cmd.add_element("name", name)
         cmd.add_element("usage_type", usage_type.value)
-        return self._send_xml_command(cmd)
-
-    def __get_configs(
-        self,
-        usage_type: UsageType,
-        *,
-        filter: Optional[str] = None,
-        filter_id: Optional[str] = None,
-        trash: Optional[bool] = None,
-        details: Optional[bool] = None,
-        families: Optional[bool] = None,
-        preferences: Optional[bool] = None,
-        tasks: Optional[bool] = None,
-    ) -> Any:
-        cmd = XmlCommand("get_configs")
-        cmd.set_attribute("usage_type", usage_type.value)
-
-        add_filter(cmd, filter, filter_id)
-
-        if trash is not None:
-            cmd.set_attribute("trash", to_bool(trash))
-
-        if details is not None:
-            cmd.set_attribute("details", to_bool(details))
-
-        if families is not None:
-            cmd.set_attribute("families", to_bool(families))
-
-        if preferences is not None:
-            cmd.set_attribute("preferences", to_bool(preferences))
-
-        if tasks is not None:
-            cmd.set_attribute("tasks", to_bool(tasks))
-
-        return self._send_xml_command(cmd)
-
-    def __get_config(
-        self,
-        config_id: str,
-        usage_type: UsageType,
-        *,
-        tasks: Optional[bool] = None,
-    ) -> Any:
-        if not config_id:
-            raise RequiredArgument(
-                function=self.get_config.__name__, argument='config_id'
-            )
-
-        cmd = XmlCommand("get_configs")
-        cmd.set_attribute("config_id", config_id)
-
-        cmd.set_attribute("usage_type", usage_type.value)
-
-        if tasks is not None:
-            cmd.set_attribute("tasks", to_bool(tasks))
-
-        # for single entity always request all details
-        cmd.set_attribute("details", "1")
-
         return self._send_xml_command(cmd)
 
     def get_feed(self, feed_type: Optional[FeedType]) -> Any:
